Move JEP 358 demo from java17 to java14
