+++ src/main/java/org/javademos/init/Java14DemoLoader.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import org.javademos.commons.IDemo;
 import org.javademos.commons.IDemoLoader;
 import org.javademos.java14.jep305.InstanceofPatternMatchingPreview;
+import org.javademos.java14.jep358.NullPointerDemo;
 import org.javademos.java14.jep359.RecordsPreviewDemo;
 import org.javademos.java14.jep361.SwitchExpressionsDemo;
 import org.javademos.java14.jep370.ForeignMemoryAccessDemo;
@@ -16,9 +17,10 @@ public class Java14DemoLoader implements IDemoLoader {
     
     @Override
     public void loadDemos(Map<Integer, IDemo> demos) {
-        demos.put(305, new InstanceofPatternMatchingPreview());
-        demos.put(359, new RecordsPreviewDemo());
-        demos.put(361, new SwitchExpressionsDemo());
-        demos.put(370, new ForeignMemoryAccessDemo());
+        demos.put(305, new InstanceofPatternMatchingPreview()); // JEP 305
+        demos.put(358, new NullPointerDemo()); // JEP 358
+        demos.put(359, new RecordsPreviewDemo()); // JEP 359
+        demos.put(361, new SwitchExpressionsDemo()); // JEP 361
+        demos.put(370, new ForeignMemoryAccessDemo()); // JEP 370
     }
 }
