+++ src/main/java/org/javademos/java14/jep358/NullPointerLevel1.java
@@ -1,4 +1,4 @@
-package org.javademos.java17.nullpointer;
+package org.javademos.java14.jep358;
 
 import lombok.Getter;
 
@@ -7,7 +7,7 @@ import lombok.Getter;
  * Here we have 4-level nested class structure and somewhere on the way
  * one element is 'null', so invoking method on it will raise 'NullPointer'.
  * 
- * @author alois.seckar@gmail.com
+ * @author Abhineshhh
  */
 public class NullPointerLevel1 {
