+++ src/main/java/org/javademos/java14/jep358/NullPointerLevel4.java
@@ -1,11 +1,11 @@
-package org.javademos.java17.nullpointer;
+package org.javademos.java14.jep358;
 
 /**
  * Simple class to show 'helpful' NullPointerExceptions from Java 14.
  * Here we have 4-level nested class structure and somewhere on the way
  * one element is 'null', so invoking method on it will raise 'NullPointer'.
  * 
- * @author alois.seckar@gmail.com
+ * @author Abhineshhh
  */
 public class NullPointerLevel4 {
